+++ android/src/main/java/com/magicleap/magicscript/ar/ModelRenderableLoaderImpl.kt
@@ -18,9 +18,9 @@ package com.magicleap.magicscript.ar
 
 import android.content.Context
 import android.net.Uri
-import android.util.TypedValue
 import com.google.ar.sceneform.assets.RenderableSource
 import com.google.ar.sceneform.rendering.ModelRenderable
+import com.magicleap.magicscript.utils.Utils
 import com.magicleap.magicscript.utils.logMessage
 
 class ModelRenderableLoaderImpl(private val context: Context) : ModelRenderableLoader {
@@ -31,7 +31,14 @@ class ModelRenderableLoaderImpl(private val context: Context) : ModelRenderableL
             result: RenderableResult<ModelRenderable>
         ) -> Unit
     ) {
-        val modelType = detectModelType(modelUri)
+        val modelType = Utils.detectModelType(modelUri, context)
+        if (modelType == ModelType.UNKNOWN) {
+            val errorMessage = "Unresolved model type"
+            logMessage(errorMessage, true)
+            resultCallback(RenderableResult.Error(Exception(errorMessage)))
+            return
+        }
+
         val builder = ModelRenderable.builder()
 
         if (modelType == ModelType.GLB) {
@@ -61,39 +68,6 @@ class ModelRenderableLoaderImpl(private val context: Context) : ModelRenderableL
                 resultCallback(RenderableResult.Error(throwable))
                 null
             }
-    }
-
-    private fun detectModelType(uri: Uri): ModelType {
-        if (uri.toString().contains("android.resource://")) { // release build
-            val resourceName = uri.lastPathSegment
-            logMessage("model res name=$resourceName")
-
-            val resourceId =
-                context.resources.getIdentifier(resourceName, "raw", context.packageName)
-            if (resourceId == 0) { // does not exists
-                return ModelType.UNKNOWN
-            }
-
-            val value = TypedValue()
-            context.resources.getValue(resourceId, value, true)
-            val resWithExtension = value.string
-            if (resWithExtension.endsWith(".glb")) {
-                return ModelType.GLB
-            }
-            if (resWithExtension.endsWith(".sfb")) {
-                return ModelType.SFB
-            }
-            return ModelType.UNKNOWN
-        } else { // localhost path
-            if (uri.toString().contains(".glb")) {
-                return ModelType.GLB
-            }
-            if (uri.toString().contains(".sfb")) {
-                return ModelType.SFB
-            }
-            return ModelType.UNKNOWN
-        }
-
     }
 
 }
