+++ android/src/main/java/com/magicleap/magicscript/utils/Utils.kt
@@ -17,11 +17,14 @@
 package com.magicleap.magicscript.utils
 
 import android.content.Context
+import android.net.Uri
+import android.util.TypedValue
 import com.google.ar.sceneform.Node
 import com.google.ar.sceneform.collision.Box
 import com.google.ar.sceneform.math.Quaternion
 import com.google.ar.sceneform.math.Vector3
 import com.google.ar.sceneform.rendering.Material
+import com.magicleap.magicscript.ar.ModelType
 import com.magicleap.magicscript.scene.nodes.base.TransformNode
 import com.magicleap.magicscript.scene.nodes.props.Bounding
 import kotlin.math.max
@@ -209,6 +212,35 @@ class Utils {
             }
         }
 
+        fun detectModelType(modelUri: Uri, context: Context): ModelType {
+            if (modelUri.toString().contains("android.resource://")) { // release build
+                val resourceName = modelUri.lastPathSegment
+                val resourceId =
+                    context.resources.getIdentifier(resourceName, "raw", context.packageName)
+                if (resourceId == 0) { // does not exists
+                    return ModelType.UNKNOWN
+                }
+
+                val value = TypedValue()
+                context.resources.getValue(resourceId, value, true)
+                val resWithExtension = value.string
+                if (resWithExtension.endsWith(".glb")) {
+                    return ModelType.GLB
+                }
+                if (resWithExtension.endsWith(".sfb")) {
+                    return ModelType.SFB
+                }
+                return ModelType.UNKNOWN
+            } else { // localhost path
+                if (modelUri.toString().contains(".glb")) {
+                    return ModelType.GLB
+                }
+                if (modelUri.toString().contains(".sfb")) {
+                    return ModelType.SFB
+                }
+                return ModelType.UNKNOWN
+            }
+        }
     }
 
 }
